+++ packages/xchain-thorchain-amm/__e2e__/thorchain-savers-e2e.ts
@@ -10,8 +10,8 @@ import {
 } from '@xchainjs/xchain-thorchain-query'
 import { assetAmount, assetToBase, baseToAsset } from '@xchainjs/xchain-util'
 
-import { Wallet } from '../src/Wallet'
 import { ThorchainAMM } from '../src/thorchain-amm'
+import { Wallet } from '../src/wallet'
 
 const thorchainQueryMainnet = new ThorchainQuery()
 const mainnetWallet = new Wallet(process.env.MAINNETPHRASE || 'you forgot to set the phrase', thorchainQueryMainnet)